Modularize RAP on profile
